backend/pyproject.toml
[project]
name = "rag-knowledge-ingestion"
version = "0.1.0"
description = "RAG Knowledge Ingestion & Vector Indexing system for Docusaurus-based technical books"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "cohere>=4.4.3",
    "qdrant-client>=1.7.0",
    "python-dotenv>=1.0.0",
]